Moving build_signaler/werkerteller to client/werkerinterface
